Don't try to install non-existent man page.
Not sure this is correct. I'm not familiar with Xcode and just edited
the file myself. In any case, xcodebuild(1) used to complain about the
non-existent file in the destroot phase, and now no longer does.

A better solution would be to write a man page :)

Found while trying to create a MacPort.